% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIrpackages-package.R
\docType{package}
\name{NVIrpackages-package}
\alias{NVIrpackages}
\alias{NVIrpackages-package}
\title{NVIrpackages: Keeps a Table of the R-Packages in NVIverse}
\description{
Keeps a table of the R-Packages in NVIverse. The table includes also meta data for the packages. The information is used in all NVIverse packages when generating vignettes and in 'NVIbatch' in functions that use a list of NVIverse packages. This table is kept in a separate package to avoid circular dependencies between packages in NVIverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NorwegianVeterinaryInstitute/NVIrpackages}
  \item Report bugs at \url{https://github.com/NorwegianVeterinaryInstitute/NVIrpackages/issues}
}

}
\author{
\strong{Maintainer}: Petter Hopp \email{Petter.Hopp@vetinst.no} (\href{https://orcid.org/0000-0002-8695-0378}{ORCID})

Other contributors:
\itemize{
  \item Norwegian Veterinary Institute [copyright holder]
}

}
\keyword{internal}
